feat(reformat): reformatted code

+++ lib/shop/data/model/response/product/product_response_model.dart
@@ -6,19 +6,19 @@ class ProductResponseModel implements JsonConverter<ProductResponseModel> {
      this.name,
      this.price,
      this.discount,
-     this.image,
-     this.inCart,
-     this.images,
-     this.description,
-     this.inFavorites,
-     this.oldPrice,
+    this.image,
+    this.inCart,
+    this.images,
+    this.description,
+    this.inFavorites,
+    this.oldPrice,
   });
 
   final int? id;
   final String? name;
-  final int? price;
-  final int? oldPrice;
-  final int? discount;
+  final double? price;
+  final double? oldPrice;
+  final double? discount;
   final String? image;
   final String? description;
   final List<String>? images;
@@ -28,20 +28,20 @@ class ProductResponseModel implements JsonConverter<ProductResponseModel> {
   @override
   ProductResponseModel fromJson(Map<String, dynamic>? jsonMap) {
     return ProductResponseModel(
-        id: jsonMap?['id'],
-        name: jsonMap?['name'],
-        price: (jsonMap?['price'] as num?)?.toInt(),
-        discount: (jsonMap?['discount'] as num?)?.toInt(),
-        image: jsonMap?['image'],
-        inCart: jsonMap?['in_cart'] ?? false,
-        images: jsonMap?['images'] != null
-            ? (jsonMap?['images'] as List).map((e) {
-          return e.toString();
-        }).toList()
-            : null,
-        description: jsonMap?['description'],
-        inFavorites: jsonMap?['in_favorites'],
-        oldPrice: (jsonMap?['old_price'] as num?)?.toInt(),
+      id: jsonMap?['id'],
+      name: jsonMap?['name'],
+      price: (jsonMap?['price'] as num?)?.toDouble(),
+      discount: (jsonMap?['discount'] as num?)?.toDouble(),
+      image: jsonMap?['image'],
+      inCart: jsonMap?['in_cart'] ?? false,
+      images: jsonMap?['images'] != null
+          ? (jsonMap?['images'] as List).map((e) {
+              return e.toString();
+            }).toList()
+          : null,
+      description: jsonMap?['description'],
+      inFavorites: jsonMap?['in_favorites'],
+      oldPrice: (jsonMap?['old_price'] as num?)?.toDouble(),
     );
   }
 }
